pyproject.toml
[tool.poetry]
name = "mas_pip"
version = "0.1.0"
description = ""
authors = ["alauddinsabari1 <aladinsabari@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7.12,<4.0"
pandas = "^1.3.1"
matplotlib = "^3.4.2"
validators = "^0.18.2"
textrazor = "^1.4.0"
requests = "^2.26.0"
beautifulsoup4 = "^4.9.3"
lxml = "^4.6.3"
simplejson = "^3.17.5"
extruct = "^0.13.0"
Wikipedia-API = "^0.5.4"
spacy-streamlit = "^1.0.2"
streamlit-aggrid = "0.2.0"
Unidecode = "^1.3.2"
google-cloud-language = "^2.3.0"
streamlit = "^1.2.0"
streamlit-lottie = "^0.0.2"
nltk = "^3.6.7"


[tool.poetry.dependencies.en_core_web_sm]
url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0.tar.gz"

[tool.poetry.dependencies.it_core_news_sm]
url = "https://github.com/explosion/spacy-models/releases/download/it_core_news_sm-3.1.0/it_core_news_sm-3.1.0.tar.gz"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
